Rename Shelf to ProductList and ShelfContainer back to Shelf
